requirements.txt
dotenv==0.9.9
matplotlib==3.10.3
MLB-StatsAPI==1.9.0
numpy==2.3.0
openpyxl==3.1.5
pandas==2.3.0
plotly==6.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
requests==2.32.4
scikit-learn==1.7.0
scipy==1.15.3
seaborn==0.13.2
espn-api==0.45.0
streamlit==1.46.0